Fix popup error
